UPSTREAM: mb/google/brya/var/agah: Update NVVDD VR PGOOD GPIO
For board revs 3 and later, the PG pin for the NVVDD VR moved from
GPP_E16 to GPP_E3. To accommodate this, the DSDT contains a Name that
this code will write the correct GPIO # to depending on the board rev,
and we'll use that instead.

BUG=b:239721380
TEST=still works on board rev 2

(cherry picked from commit bcc3059d83b7503a3247690a895a05127102d414)

Original-Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Change-Id: I11aec6069da8e086789419303871c6d0f5fb29af
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/66806
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Eric Lai <eric_lai@quanta.corp-partner.google.com>
Original-Reviewed-by: Ivy Jian <ivy.jian@quanta.corp-partner.google.com>
GitOrigin-RevId: bcc3059d83b7503a3247690a895a05127102d414
Change-Id: I8ef7c8bba7677d2ed5668ffb159d42ca52c68d78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3841926
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Kangheui Won <khwon@chromium.org>
Commit-Queue: Kangheui Won <khwon@chromium.org>