UPSTREAM: mb/system76: Set SMBIOS wakeup type to power switch
Windows hardware tests require this field not be "Reserved".

The System76 EC firmware does not report the wake type, so it is not
possible to know if the system was powered on from the power switch or
Wake-on-LAN. In the case WoL is used, this will report the wrong value.

(cherry picked from commit e086d149ce9de5d29f1b543a0e71ff61babb0144)

Original-Change-Id: I4653c6bce2a5f0a88281fc810df5646e44f90674
Original-Signed-off-by: Tim Crawford <tcrawford@system76.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/66837
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Jeremy Soller <jeremy@system76.com>
GitOrigin-RevId: e086d149ce9de5d29f1b543a0e71ff61babb0144
Change-Id: Ifd299f23b3aae2a7a110ffb5fa4c73a620329156
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3948931
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Rob Barnes <robbarnes@google.com>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Commit-Queue: Rob Barnes <robbarnes@google.com>