UPSTREAM: util/lint/lint: Add -I option to invert test results
To test the linters, we want to invert the results so that any test that
passes shows up as a failure. This will allow us to verify that all of
the linters are working correctly.

This will be tested nightly as well as on changes to the lint tools.

(cherry picked from commit 38bbff47a718146887692e36c950094d0801a46c)

Original-Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Change-Id: Ia8024c6ab0c91fd9f630f37dc802ed3bc6b4608c
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/67193
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
GitOrigin-RevId: 38bbff47a718146887692e36c950094d0801a46c
Change-Id: Ia2f8d2610d120d18bc28bcc2fa8d9addd231cde4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3882532
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Reviewed-by: Reka Norman <rekanorman@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>