UPSTREAM: Add SBOM (Software Bill of Materials) Generation
Firmware is typically delivered as one large binary image that gets
flashed. Since this final image consists of binaries and data from
a vast number of different people and companies, it's hard to
determine what all the small parts included in it are. The goal of
the software bill of materials (SBOM) is to take a firmware image
and make it easy to find out what it consists of and where those
pieces came from. Basically, this answers the question, who supplied
the code that's running on my system right now? For example, buyers
of a system can use an SBOM to perform an automated vulnerability
check or license analysis, both of which can be used to evaluate
risk in a product. Furthermore, one can quickly check to see if the
firmware is subject to a new vulnerability included in one of the
software parts (with the specified version) of the firmware.
Further reference:
https://web.archive.org/web/20220310104905/https://blogs.gnome.org/hughsie/2022/03/10/firmware-software-bill-of-materials/

- Add Makefile.inc to generate and build coswid tags
- Add templates for most payloads, coreboot, intel-microcode,
  amd-microcode. intel FSP-S/M/T, EC, BIOS_ACM, SINIT_ACM,
  intel ME and compiler (gcc,clang,other)
- Add Kconfig entries to optionally supply a path to CoSWID tags
  instead of using the default CoSWID tags
- Add CBFS entry called SBOM to each build via Makefile.inc
- Add goswid utility tool to generate SBOM data

(cherry picked from commit 78bddb8a51564ac861ebf86cf38ac8600f452d41)

Original-Signed-off-by: Maximilian Brune <maximilian.brune@9elements.com>
Original-Change-Id: Icb7481d4903f95d200eddbfed7728fbec51819d0
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/63639
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martin.roth@amd.corp-partner.google.com>
CopyBot-Skipped-File: .gitmodules
GitOrigin-RevId: 1d7a9debf241f9649a40ebc367204bac0a86a67e
Change-Id: I7f87780678dad4bb187ed69bd2aa0ba5738699e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3853707
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Kangheui Won <khwon@chromium.org>
Reviewed-by: Kangheui Won <khwon@chromium.org>