UPSTREAM: soc/intel/spi: Move BIOS flash SPI controllers to fast SPI driver
There are two classes of SPI controllers on Intel chipsets:
 * generic usable SPI controllers
 * SPI controller hosting the BIOS flash (fast SPI controller)

While the first class can be used for generic peripheral attachment the
second class mostly controls the BIOS flash and a TPM device (if
enabled). The generic SPI driver is not fully applicable to the fast SPI
controller. In addition, the fast SPI controller reports the reserved
MMIO range used for the BIOS flash mapping so that the OS is aware of
this range.

This patch moves the fast SPI controller of all known SoCs to the
fast SPI driver in common code. In addition, the PCI device for the
fast SPI controller is removed from the function 'spi_soc_devfn_to_bus'
as this is a callback of the generic SPI driver.

(cherry picked from commit 777099046f88f96921e21ad2e0a58b2fe66b25aa)

Original-Change-Id: Ia881c1d274acdcf7f042dd8284048a7dd018a84b
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/67173
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
GitOrigin-RevId: 777099046f88f96921e21ad2e0a58b2fe66b25aa
Change-Id: Iefeef011098c18effcda5dd8f5ad4d0c8d5e6ba6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3916171
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>