UPSTREAM: src/vendorcode: Add a README.md file
This README.md file specifies the expectations and agreements implied by
contributing code to the coreboot src/vendorcode directory.

Licenses, structure, formatting and the like are all the responsibility
of the contributors, however as the code is a part of the coreboot
codebase, members of the coreboot community are allowed to modify the
code to the extent deemed necessary.

(cherry picked from commit 5962b8a91b809b613186227c00d17aed66d5c031)

Original-Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Change-Id: Ib7e2aedce9383158d03be342cb10ae18d85146ae
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/67485
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: 5962b8a91b809b613186227c00d17aed66d5c031
Change-Id: Ia99c30bdf4778a636f80060de71597ef53263879
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3978586
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>