UPSTREAM: soc/intel/mtl: Fix GPIO group pad base for ACPI
This patch fixes MeteorLake GPIO PINCTRL entries as per 5.15
kernel pintrl driver:
https://source.chromium.org/chromiumos/chromiumos/codesearch/+/main:src/third_party/kernel/v5.15/drivers/pinctrl/intel/pinctrl-meteorlake.c

In order to support using ACPI GPIOs it is necessary for coreboot
to be compatible with this implementation.  The GPIO groups that
are usable by the  OS are declared with a pad base which is then
used to compute the number for ACPI GPIOs.

BUG=b:232573696
TEST=Tested on Google Rex board. After this change, driver rt5682s
is able to claim pinctrl IRQ 358 corresponding to GPP_B06.

(cherry picked from commit 715c17a750e5377df353eb93e4ca5fda395e4225)

Original-Signed-off-by: Kapil Porwal <kapilporwal@google.com>
Original-Change-Id: Icabbe9e125ee9efaf0eef4c4cdc8be9f734aa703
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/67565
Original-Reviewed-by: Eric Lai <eric_lai@quanta.corp-partner.google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Ivy Jian <ivy.jian@quanta.corp-partner.google.com>
Original-Reviewed-by: Subrata Banik <subratabanik@google.com>
Original-Reviewed-by: Tarun Tuli <taruntuli@google.com>
GitOrigin-RevId: 715c17a750e5377df353eb93e4ca5fda395e4225
Change-Id: I9b98b8805c11cf19fc30c46fbccbd16dc5238383
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3895620
Tested-by: Subrata Banik <subratabanik@chromium.org>
Reviewed-by: Subrata Banik <subratabanik@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Subrata Banik <subratabanik@chromium.org>