UPSTREAM: drivers/mrc_cache: Compare hashes instead of full data
The current MRC cache update process is slow (28 ms on nissa), because
cbmem is not cached in romstage. Specifically, the new MRC data returned
by the FSP is stored in the FSP reserved memory in cbmem, so operations
on the new data (computing the checksum, comparing to the old data) are
slow.

Replace the data checksum in the MRC header with a hash, and compare
hashes instead of comparing the full data. This has two benefits:
1. The xxhash function is faster than computing an IP checksum (4 ms vs
   14 ms on uncached data on nissa).
2. There's no need to memcmp() the full MRC data, which takes 14 ms on
   nissa.

Before:
550:starting to load ChromeOS VPD                     867,930 (4,664)
  3:after RAM initialization                          896,020 (28,090)
  4:end of romstage                                   906,274 (10,254)

After:
550:starting to load ChromeOS VPD                     864,820 (4,649)
  3:after RAM initialization                          869,652 (4,831)
  4:end of romstage                                   879,909 (10,257)

BUG=b:242667207
TEST=Check that MRC caching still works as expected on nissa. Corrupt
the MRC cache and check that memory is retrained.

(cherry picked from commit ee63b44c47cbb8c81ce0e173bbc12129c3edbe71)

Cq-Depend: chromium:3939403
Original-Change-Id: I1b7848d1d05e555b61e0f1cb605550dfe3449c6d
Original-Signed-off-by: Reka Norman <rekanorman@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/67670
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Reviewed-by: Eric Lai <eric_lai@quanta.corp-partner.google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: ee63b44c47cbb8c81ce0e173bbc12129c3edbe71
Change-Id: If8716553ace0978b936b4e3c4b475b0c5a6d0cc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3914920
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Tested-by: Reka Norman <rekanorman@chromium.org>