UPSTREAM: mb/google/brya/var/banshee: Remove privacy_gpio
On Banshee, when the privacy switch is toggled the camera is
disconnected. Which means that we will never be able to tell the user
that the privacy switch is enabled when the camera is on, making the
virtual control unusable.

Remove the description.

BUG=b:248219472
BRANCH=firmware-brya-14505.B
TEST=none

(cherry picked from commit e7b54bfc36fdb3c9d74d0718e47dd656ab0f22f0)

Original-Change-Id: I1a241bd889c0c1aae039510a0620748b2f7a6806
Original-Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/67773
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
GitOrigin-RevId: e7b54bfc36fdb3c9d74d0718e47dd656ab0f22f0
Change-Id: Icbadd4768ab2fbcc327dc6e930ff23244223d475
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3918688
Commit-Queue: Nick Vaccaro <nvaccaro@google.com>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Ricardo Ribalda <ribalda@chromium.org>