UPSTREAM: util/amdfwtool/data_parse: fix PMU subprogram/instance ID handling
The parsing of the PMU binary subprogram and instance numbers only
worked correctly for the cases where the ID in the name in the fw.cfg
file was between 0 and 9, but returned wrong results if it was between a
and f. Switch to using strtol with a base of 16 instead of subtracting
the char '0' from the char in the filename in
find_register_fw_filename_bios_dir to fix this.

(cherry picked from commit 3dfb48533484b8ba6ba1b1c1a39fcfee9491989c)

Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: Ic5fd41daf9f26d11c1f86375387c1d7beac04124
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/67927
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Matt DeVillier <matt.devillier@gmail.com>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
GitOrigin-RevId: 3dfb48533484b8ba6ba1b1c1a39fcfee9491989c
Change-Id: Ib34dbb7cda35fa2bee0aff3ca7541c6253fdea96
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3931641
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>