UPSTREAM: soc/amd/common: Remove buildtime error for unknown cpu
This is not critical functionality and doesn't need a build-time error.
Having it as a build time error causes a chicken & egg issue where
the chipset needs to be added before it can be added to this file, but
the header file fails the build because the chipset is unknown.

It's not practical to exclude these files from the new platform builds
because the PSP functionality is thoroughly embedded into the coreboot
structure.

(cherry picked from commit f225d761ba1ae6dea878c33668b590fe94293fe5)

Original-Signed-off-by: Martin Roth <martin.roth@amd.corp-partner.google.com>
Original-Change-Id: Ib02bbe1f9ffb343e1ff7c2bfdc45e7edffe7aaed
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/68245
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Reviewed-by: Fred Reitberger <reitbergerfred@gmail.com>
Original-Reviewed-by: Paul Menzel <paulepanter@mailbox.org>
GitOrigin-RevId: f225d761ba1ae6dea878c33668b590fe94293fe5
Change-Id: I5e712dbd8bce4be1844d5d6db47b9abbad3a2f7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3943205
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Rob Barnes <robbarnes@google.com>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Tested-by: Rob Barnes <robbarnes@google.com>