UPSTREAM: configs: Add skyrim config with binaries
We've seen failures because the binaries were not being built into the
image.  In particular, the APCB is modified by the coreboot build
process, so if the APCB isn't built correctly to support the correct
number of SPDs, the build can fail.

The mendocino FSP binaries are not yet pushed, so the build is currently
pointing at the cezanne binaries.  The mendocino FSP will be pushed when
the mendocino chips are released for sale.

(cherry picked from commit 08e7df9d5d93225d263b29c38ba7d4d9945f6dfc)

Original-Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Change-Id: I31d11c5327416f4339930373c447531ae9f79d28
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/68320
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Jon Murphy <jpmurphy@google.com>
GitOrigin-RevId: 08e7df9d5d93225d263b29c38ba7d4d9945f6dfc
Change-Id: If91ea79bd22372a3c62a49b64dd97528ac165c34
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3979740
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>