UPSTREAM: soc/amd/stoneyridge: add chipset devicetrees
Add chipset devicetrees for Stoneyridge and Carrizo, which is also
supported by the Stoneyridge code, but has more external PCIe ports and
devices. The mainboard's devicetrees will be changed to use the aliases
defined in the chipset devicetree in follow-up patches. This is a
preparation to statically assign the ops for the internal devices
statically in the SoC devicetree instead of dynamically adding them in
ramstage.

BKDG #55072 Rev 3.04 was used to check the PCI devices and functions and
the MMIO addresses.

(cherry picked from commit b68e22409d8e22e097193bd26cb31213c7030db7)

Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: Ia45260b1168ed1d99993adfb98475da5b5c90d11
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/68316
Original-Reviewed-by: Matt DeVillier <matt.devillier@amd.corp-partner.google.com>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: b68e22409d8e22e097193bd26cb31213c7030db7
Change-Id: Ie6f0a157c574aa76d8675a284361a1e5b911930c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3955586
Reviewed-by: Reka Norman <rekanorman@chromium.org>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>