UPSTREAM: soc/intel/eklhartlake: Provide an option to disable the L1 prefetcher
Depending on the real workload that is executed on the system the L1
prefetcher might be too aggressive and will populate the L1 cache ahead
with data that is not really needed. In the end, this will result in a
higher cache miss rate thus slowing down the real application.

This patch provides a devicetree option to disable the L1 prefetcher if
needed. This can be requested on mainboard level if needed.

(cherry picked from commit d03e896b57ea1ae2fea640937df2d57b684ce629)

Original-Change-Id: I3fc8fb79c42c298a20928ae4912ee23916463038
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/68667
Original-Reviewed-by: Lean Sheng Tan <sheng.tan@9elements.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: d03e896b57ea1ae2fea640937df2d57b684ce629
Change-Id: I25e01d776ee297bd6a0412ea0772332fc6691436
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3979758
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>