UPSTREAM: ifdtool: Determine max regions from IFD
IFDv1 always has 8 regions, while IFDv2 always has 16 regions.

It's platform specific which regions are used or are reserved.
The 'SPI programming guide' as the name says is a guide only,
not a specification what the hardware actually does.
The best to do is not to rely on the guide, but detect how many
regions are present in the IFD and expose them all.

Very early IFDv2 chipsets, sometimes unofficially referred to as
IFDv1.5 platforms, only have 8 regions. To not corrupt the IFD when
operating on an IFDv1.5 detect how much space is actually present
in the IFD.

Fixes IFD corruption on Wellsburg/Lynxpoint when writing a new
flash layout.

(cherry picked from commit 98ecaa4a55badd09fe85a56d28d3c1793b52b963)

Original-Change-Id: I0e3f23ec580b8b8402eb1bf165e3995c8db633f1
Original-Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/68780
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Lean Sheng Tan <sheng.tan@9elements.com>
Original-Reviewed-by: Christian Walter <christian.walter@9elements.com>
GitOrigin-RevId: 98ecaa4a55badd09fe85a56d28d3c1793b52b963
Change-Id: I151a932343f909badba81cf9ad69082de6a1f834
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4155689
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Commit-Queue: Rob Barnes <robbarnes@google.com>