UPSTREAM: mb/google/kahlee/liara/devicetree: move Raydium touchscreen to baseboard
Move the Raydium touchscreen to the baseboard devicetree. Since only the
liara variant uses a level IRQ as I2C devices are supposed to, all other
board variants still override this to use an edge IRQ which were added
as a workaround to make the touchscreen work on the other devices. Right
now it's unclear to me if that edge IRQ workaround was only needed
temporarily and can now be removed, so I'll keep it as it was for now.
If this turns out to be no longer needed on the other variants, the
overrides can be dropped in the future.

(cherry picked from commit 6990cb29ab02d2c706011415145ff4dab22e0703)

Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: Ic621c1a5856e9e280a25b0668010a1ee5bbb61e4
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/68770
Original-Reviewed-by: Matt DeVillier <matt.devillier@amd.corp-partner.google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Paul Menzel <paulepanter@mailbox.org>
GitOrigin-RevId: 6990cb29ab02d2c706011415145ff4dab22e0703
Change-Id: I5df2f94954590fe3a155cd012b5481a611f5360f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3979769
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>