UPSTREAM: mb/google/kahlee/mainboard: rewrite IRQ mapping handling
Rewrite the Kahlee IRQ mapping handling to be in line with the newer AMD
SoCs to allow making the largest part of the corresponding code common
for all AMD SoCs in the coreboot tree.

The PIC-mode IRQ numbers for both PIRQ_ASF and PIRQ_SDIO were 0 in the
data tables which is the PIT IRQ which looks very wrong to me, so it was
changed to PIRQ_NC. Since the ASF and likely also the SDIO controller
are unused, this shouldn't change runtime behavior. The data tables also
had non 0 and non 0x1f entries in the following locations the internal
BKDG #55072 revision 3.04 describes as unused: 0x31, 0x33, 0x35-0x37,
0x40, 0x50-0x53. The entry at 0x32 is also non 0 and non 0x1f and the
description in the BKDG says that it controls the IRQ mapping of another
internal PCI device, but that PCI device doesn't exist in the SoC.

TEST=No obvious IRQ-related breakage on google/liara

(cherry picked from commit 7e247a3fa8cd2e039200b497ee67daff16953426)

Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Tested-by: Matt DeVillier <matt.devillier@amd.corp-partner.google.com>
Original-Change-Id: I9b3bfca33d88ef3989b63f4fe6c301e0e485b7e9
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/68851
Original-Reviewed-by: Matt DeVillier <matt.devillier@amd.corp-partner.google.com>
Original-Reviewed-by: Martin Roth <martin.roth@amd.corp-partner.google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: 7e247a3fa8cd2e039200b497ee67daff16953426
Change-Id: I5758e844f0fd0890bc990f1f80c8161180cfc309
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4030696
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>