UPSTREAM: device/dram: Add kconfig options for memory types
Currently, we're building support for all memory types into every board,
and letting the linker remove anything that isn't needed. This is okay,
but it'd be nice to be able to build in just what's actually needed.

This change adds options to specify both what is used and what is not.
By doing it that way, the default values don't change, but platforms can
start removing support for memory types that are not needed.  When all
platforms (SoCs, CPUs and/or Northbridge chips) specify what memory
types they support, the defaults on the options to use a particular
memory type can be set to no, and the options not to use a memory type
can be removed.

(cherry picked from commit a666af7b015e7625c3d1ae679d50c69190c6cd5d)

Original-Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Change-Id: I07c98a702e0d67c5ad7bd9b8a4ff24c9288ab569
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/68992
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Elyes Haouas <ehaouas@noos.fr>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Eric Lai <eric_lai@quanta.corp-partner.google.com>
GitOrigin-RevId: a666af7b015e7625c3d1ae679d50c69190c6cd5d
Change-Id: I8ab7f63e0754e429727f3b2ee1063ffebaa194e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4005986
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Patrick Georgi <pgeorgi@google.com>
Reviewed-by: Kangheui Won <khwon@chromium.org>