UPSTREAM: src/device + util/sconfig: Introduce new device 'mdio'
This patch extends the available device paths with a new device 'mdio'.
MDIO is the 'Management Data Input/Output' called interface which is
used to access an Ethernet PHY behind a MAC to change settings. The real
payload data path is not handled by this interface.

To address the PHY correctly on the MDIO bus, there is a 5 bit address
needed, which often can be configured via pins on the mainboard.
Therefore, the new introduced device has an 'addr' field to define its
address. If one wants to use a MDIO device in devicetree, the syntax is
straight forward (example):
	device mdio 0x2 on end

As the MDIO interface is driven by the MAC, most likely this MDIO device
will be hooked in as a child device of the (PCI attached) MAC device.

With the new introduced ops_mdio a new interface is added to provide an
API for read and write access over MDIO.

(cherry picked from commit 67f63e768d8860ebc6bae5987e2d928efabcf7c4)

Original-Change-Id: I6691f92c4233bc30afc9029840b06f74bb1eb4b2
Original-Signed-off-by: Mario Scheithauer <mario.scheithauer@siemens.com>
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/69382
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: 67f63e768d8860ebc6bae5987e2d928efabcf7c4
Change-Id: I361794782afb666e4c5e0d2f3d3389ec3da68957
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4053416
Reviewed-by: Rob Barnes <robbarnes@google.com>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Tested-by: Rob Barnes <robbarnes@google.com>
Commit-Queue: Rob Barnes <robbarnes@google.com>