UPSTREAM: cpu/x86/smm/module_loader: Fix ASEG loading
This code was never tested with SSE enabled. Now qemu enables it and
FX_SAVE encroaches on the save states. Without SSE enabled the handler
just happened to be aligned downwards enough to have the save states
fit. With SSE enabled that's not the case. The proper fix is to give the
code setting up stubs the right base address, which is the same as for
the TSEG codepath.

(cherry picked from commit f874fc2717cf1a7b06bfb73f3de3698f1e1a9244)

Original-Change-Id: I45355efb274c6ddd09a6fb57743d2f6a5b53d209
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/69233
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Kyösti Mälkki <kyosti.malkki@gmail.com>
GitOrigin-RevId: f874fc2717cf1a7b06bfb73f3de3698f1e1a9244
Change-Id: I3ae71b57b933c03215cd10bc6eb705f6765b2a29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4021094
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>