UPSTREAM: cpu/intel/speedstep: Have nb and sb code provide c5/c6/slfm
C5, C6 and slfm depend on the southbridge and the northbridge to be able
to provide this functionality, with some just lacking the possibility to
do so. Move the devicetree configuration to the southbridge.

This removes the need for a magic lapic in the devicetree.

(cherry picked from commit 98c92570d9bb363740ae1b2cbbefc3c0f2404cb4)

Original-Change-Id: I4a9b1e684a7927259adae9b1d42a67e907722109
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/69297
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
GitOrigin-RevId: 98c92570d9bb363740ae1b2cbbefc3c0f2404cb4
Change-Id: I588a739c9f8c3bb19d437b30fe3317aa953521a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4080638
Commit-Queue: Kangheui Won <khwon@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Kangheui Won <khwon@chromium.org>