UPSTREAM: util/kconfig/README.md: Add notes about adding a new quilt patch
The patches for kconfig need to be in a format compatible with the quilt
tool, and usually also contain a header with some additional info like
the git commit. This header is in the same format as patches produced by
`git format-patch`, but the diff style git uses is incompatible with
quilt and there does not seem to be a straightforward way to format the
diff section to work.

Add some documentation for a method I found to go from a git commit to a
quilt compatible patch with git headers.

(cherry picked from commit 9eab93168d26b6a459bf59a7a4f0344ada9872b3)

Original-Change-Id: I7a8bbe41e0864be1d28116742b6b8b3fc440cc31
Original-Signed-off-by: Nicholas Chin <nic.c3.14@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/69458
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
GitOrigin-RevId: 9eab93168d26b6a459bf59a7a4f0344ada9872b3
Change-Id: Id96350b458bb43e720b6b879228f7975f095771c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4075523
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Kangheui Won <khwon@chromium.org>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Commit-Queue: Kangheui Won <khwon@chromium.org>