UPSTREAM: Revert "src/arch/x86: Use core apic id to get cpu_index()"
This reverts commit 095c931cf12924da9011b47aa64f4a6f11d89f13.

Previously cpu_info() was implemented with a struct on top of an
aligned stack. As FSP changed the stack value cpu_info() could not be
used in FSP context (which PPI is). Now cpu_info() uses GDT segments,
which FSP does not touch so it can be used.

This also exports cpu_infos from cpu.c as it's a convenient way to get
the struct device * for a certain index.

TESTED on aldrvp: FSP-S works and is able to run code on APs.

(cherry picked from commit cc22607dbfbab0c9ce42c071b5b3c4a304845313)

Original-Change-Id: I3a40156ba275b572d7d1913d8c17c24b4c8f6d78
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/69509
Original-Reviewed-by: Subrata Banik <subratabanik@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: cc22607dbfbab0c9ce42c071b5b3c4a304845313
Change-Id: Ia27aac835ab6b168940f96843fd3016505a36b7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4066574
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Rob Barnes <robbarnes@google.com>
Reviewed-by: Rob Barnes <robbarnes@google.com>