UPSTREAM: commonlib/fsp_relocate.c: Fix cbfstool build on 32-bit host
On a 32-bit host, uintptr_t is defined as 'unsigned int' instead of
'unsigend long int' like on a 64-bit host. When cbfstool is built on a
32-bit host, the printk format specifier '%lx' expects a 'long int'
while new_addr is of type 'uintptr_t', aka 'unsigned int'.
This in the end leads to a build error.

To fix this and make it build on both, 32- and 64-bit hosts, use PRIxPTR
as the format specifier. This macro will be resolved at compile time in
the right way on both, 32- and 64-bit hosts.

Test=Build cbfstool on 32- and 64-bit hosts.

Original-Change-Id: Ia917d2ed31778f3a29c0a6c7368f74c15319b099
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/69682
Original-Reviewed-by: Mario Scheithauer <mario.scheithauer@siemens.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Change-Id: Ie71e3f92c3a4c24a6d51fe67a1d6b6713bb3fe95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4113207
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Tested-by: Reka Norman <rekanorman@chromium.org>