UPSTREAM: build: List all Kconfigs in CBFS `config` file, compress it
The coreboot build system automatically adds a `config` file to CBFS
that lists the exact Kconfig configuration that this image was built
with. This is useful to reproduce a build after the fact or to check
whether support for a specific feature is enabled in the image.

However, the file is currently generated using the `savedefconfig`
command to Kconfig, which generates the minimal .config file that is
needed to produce the required config in a coreboot build. This is fine
for reproduction, but bad when you want to check if a certain config was
enabled, since many configs get enabled by default or pulled in through
another config's `select` statement and thus don't show up in the
defconfig.

This patch tries to fix that second use case by instead including the
full .config instead. In order to save some space, we can remove all
comments (e.g. `# CONFIG_XXX is not set`) from the file, which still
makes it easy to test for a specific config (if it's in the file you can
extract the right value, if not you can assume it was set to `n`). We
can also LZMA compress it since this file is never read by firmware
itself and only intended for later re-extraction via cbfstool, which
always has LZMA support included.

On a sample Trogdor device the existing (uncompressed) `config` file
takes up 519 bytes in CBFS, whereas the new (compressed) file after this
patch will take up 1832 bytes -- still a small amount that should
hopefully not break the bank for anyone.

(cherry picked from commit 4924cdb9ace88336cc37506a745ebef210f20767)

Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I5259ec6f932cdc5780b8843f46dd476da9d19728
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/69710
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Yu-Ping Wu <yupingso@google.com>
Original-Reviewed-by: Jakub Czapiga <jacz@semihalf.com>
Original-Reviewed-by: Martin Roth <martin.roth@amd.corp-partner.google.com>
GitOrigin-RevId: 4924cdb9ace88336cc37506a745ebef210f20767
Change-Id: I31e280abb97809e086e763756c53074b95834c92
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4040731
Tested-by: Nick Vaccaro <nvaccaro@google.com>
Commit-Queue: Nick Vaccaro <nvaccaro@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>